removed generics added quote detection to Shell
git-svn-id: svn://localhost/trunk@980 f2a0fb1d-5e3f-4af0-b388-fe59793aa704

+++ src/main/java/klon/Compiler.java
@@ -155,7 +155,7 @@ public class Compiler extends KlonAnalyzer implements KlonConstants {
       String image = node.getImage();
       if (image.startsWith("0x") || image.startsWith(("0X"))) {
         node.addValue(root.getSlot("Number").clone(
-            (double) Integer.parseInt(image.substring(2), 16)));
+            Integer.parseInt(image.substring(2), 16)));
       } else {
         node.addValue(root.getSlot("Number").clone(Double.parseDouble(image)));
       }
